+++ src/enemy/spin_tile.cpp
@@ -9,41 +9,41 @@ SpinTile::SpinTile(const enemy& other, bool newScriptUID, bool clearParentUID):
 	if(other.scrmem)
 	{
 		alloc_scriptmem();
-		memcpy(scrmem->stack, other.scrmem->stack, MAX_SCRIPT_REGISTERS*sizeof(long));
+		memcpy(scrmem->stack, other.scrmem->stack, MAX_SCRIPT_REGISTERS*sizeof(int32_t));
 		scrmem->scriptData = other.scrmem->scriptData;
 	}
 	else
 		scrmem = nullptr;
 
-	for(int i = 0; i < edefLAST255; ++i)
+	for(int32_t i = 0; i < edefLAST255; ++i)
 		defense[i] = other.defense[i];
-	for(int q = 0; q < 10; ++q)
+	for(int32_t q = 0; q < 10; ++q)
 		frozenmisc[q] = other.frozenmisc[q];
-	for(int q = 0; q < NUM_HIT_TYPES_USED; +q)
+	for(int32_t q = 0; q < NUM_HIT_TYPES_USED; +q)
 		hitby[q] = other.hitby[q];
 
 	if(newScriptUID)
 		script_UID = FFCore.GetScriptObjectUID(UID_TYPE_NPC); //This is used by child npcs.
 	if(clearParentUID)
 		parent_script_UID = 0;
-	for(int q = 0; q < 32; ++q)
+	for(int32_t q = 0; q < 32; ++q)
 		movement[q] = other.movement[q];
-	for(int q = 0; q < 32; ++q)
+	for(int32_t q = 0; q < 32; ++q)
 		new_weapon[q] = other.new_weapon[q];
 
-	for(int q = 0; q < 8; ++q)
+	for(int32_t q = 0; q < 8; ++q)
 	{
 		initD[q] = other.initD[q];
 		weap_initiald[q] = other.weap_initiald[q];
 	}
-	for(int q = 0; q < 2; ++q)
+	for(int32_t q = 0; q < 2; ++q)
 	{
 		initA[q] = other.initA[q];
 		weap_initiala[q] = other.weap_initiala[q];
 	}
 }
 
-SpinTile::SpinTile(zfix x, zfix y, int id, int comboTile):
+SpinTile::SpinTile(zfix x, zfix y, int32_t id, int32_t comboTile):
 	enemy(x, y, id&0xFFF, 0)
 {
 	if(comboTile > 0)
@@ -90,7 +90,7 @@ SpinTile::SpinTile(zfix x, zfix y, int id, int comboTile):
 		zofs = d->zofs;
 }
 
-bool SpinTile::animate(int index)
+bool SpinTile::animate(int32_t index)
 {
 	if(fallclk || drownclk)
 		return enemy::animate(index);
