+++ src/util.h
@@ -10,7 +10,7 @@
 #ifndef _WIN32
 #include <stdlib.h>
 #else
-typedef int mode_t;
+typedef int32_t mode_t;
 #endif
 namespace util
 {
@@ -25,14 +25,14 @@ namespace util
 	void regulate_path(char* buf);
 	void regulate_path(std::string& buf);
 	bool create_path(const char* path);
-	char* zc_itoa(int value, char* str, int base = 10);
-	long long zc_atoi64(const char *str);
-	long long zc_xtoi64(const char *hexstr);
-	int zc_xtoi(const char *hexstr);
-	long ffparse2(const char *string);
-	long ffparseX(const char *string);
-	int xtoi(char *hexstr);
-	int zc_chmod(const char* path, mode_t mode);
+	char* zc_itoa(int32_t value, char* str, int32_t base = 10);
+	int64_t zc_atoi64(const char *str);
+	int64_t zc_xtoi64(const char *hexstr);
+	int32_t zc_xtoi(const char *hexstr);
+	int32_t ffparse2(const char *string);
+	int32_t ffparseX(const char *string);
+	int32_t xtoi(char *hexstr);
+	int32_t zc_chmod(const char* path, mode_t mode);
 	bool checkPath(const char* path, const bool is_dir);
 	void safe_al_trace(const char* str);
 	direction angleToDir4(double angle);
@@ -40,7 +40,7 @@ namespace util
 	direction dir4To(int fromX, int fromY, int toX, int toY);
 	direction dir8To(int fromX, int fromY, int toX, int toY);
 }
-int vbound(int val, int low, int high);
+int32_t vbound(int32_t val, int32_t low, int32_t high);
 double vbound(double val, double low, double high);
 
 //#define vbound(x,low,high) (x<low?low:(x>high?high:x))
